removes extra node source file. Moves optimizer test code to tests
